+++ src/Caffeine/Classes/Models/KeyboardShortcutsName+Caffeine.swift
@@ -1,0 +1,10 @@
+//
+//  KeyboardShortcutsName+Caffeine.swift
+//  Caffeine
+//
+
+import KeyboardShortcuts
+
+extension KeyboardShortcuts.Name {
+    static let toggleActive = Self("toggleActive")
+}